Factor out (potentially) common setup and fix connection closing
