+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/persistence/CordaPersistence.kt
@@ -93,7 +93,8 @@ class CordaPersistence(
         val jdbcUrl: String,
         cacheFactory: NamedCacheFactory,
         attributeConverters: Collection<AttributeConverter<*, *>> = emptySet(),
-        customClassLoader: ClassLoader? = null
+        customClassLoader: ClassLoader? = null,
+        val closeConnection: Boolean = true
 ) : Closeable {
     companion object {
         private val log = contextLogger()
@@ -244,7 +245,7 @@ class CordaPersistence(
                     throw e
                 }
             } finally {
-                quietly(transaction::close)
+                quietly { transaction.close() }
             }
         }
     }
