+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/persistence/DatabaseTransaction.kt
@@ -96,8 +96,10 @@ class DatabaseTransaction(
         if (sessionDelegate.isInitialized() && session.isOpen) {
             session.close()
         }
-        // TODO: add a way of not closing the transaction.
-        //connection.close()
+
+        if (database.closeConnection) {
+            connection.close()
+        }
         contextTransactionOrNull = outerTransaction
         if (outerTransaction == null) {
             synchronized(this) {
